#16 Set up automation
Create .github/workflows/test_workflow.yml defining the steps required to start the tests on the library using the containers from #2. These tests are triggered when a pull request is created on master and when a commit is pushed on a branch starting with "test", allowing developers to trigger the tests on Github without creating a pull request.